Removing Redundant Code in Read Mail
Unnecessarily repeated codes
