+++ src/UI/Components/Mail/ReadMail.html
@@ -1,0 +1,69 @@
+<div id="ReadMail">
+	<div class="body" data-background="basic_interface/maillist3_bg.bmp">
+		<div class="titlebar">
+			<div class="left">
+				<span class="text" id="title" data-text="1027">Read Mail</span>
+			</div>
+			<div class="right">
+				<button class="base close" data-background="basic_interface/close2.bmp" data-hover="basic_interface/close2_a.bmp"></button>
+			</div>
+	
+		</div>
+
+		<div class="clear">
+
+			<section class="container flex_grid">
+
+				<div class="title" >
+					<div >
+						<span class="text_title">
+							
+						</span>
+					</div>					
+				</div>
+
+				<div class="name_sender" >
+					<div >
+						<span class="text_sender">
+							
+						</span>
+					</div>
+				</div>
+
+				<div class="mail_body" >
+					<div >
+						<textarea class="textarea_mail" maxlength="198" disabled></textarea>
+					</div>					
+				</div>
+
+				<div class="zeny_item">					
+					<div class="zeny_item_container">
+						
+						<div class="zeny_item_infor_box"></div>
+						<div class="zeny_item_infor event_add_cursor">
+							<div class="block_zeny">
+								<input class="input_zeny_amt event_add_cursor" type="text" value="" maxlength="9" disabled/>
+							</div>
+							<!-- item -->
+							<div class="block_item">
+								<div class="container_item" >
+									<div class="overlay"></div>
+								</div>
+							</div>
+						</div>
+					</div>    
+				</div>
+				<div class="btn_return_reply_remove">
+					<div class="container_return_reply_remove">						
+						<button class="read_mail_btn" id="read_mail_return" data-background="basic_interface/return.bmp" data-hover="basic_interface/return_a.bmp"></button>
+						<button class="read_mail_btn" id="read_mail_remail" data-background="basic_interface/remail.bmp" data-hover="basic_interface/remail_a.bmp"></button>
+						<button class="read_mail_btn" id="read_mail_del" data-background="basic_interface/del.bmp" data-hover="basic_interface/del_a.bmp"></button>
+					</div>
+					
+				</div>
+
+			</section>
+		</div>
+	</div>
+	
+</div>
